:lipstick: Fix general android styles
