+++ src/components/navigation/HeaderSearchBar/styles.js
@@ -1,8 +1,16 @@
 import { colors, fonts, boxes } from 'constants/styleGuide';
+import { deviceType } from 'utilities/device';
+
+const type = deviceType();
+
+let normalMarginTop = type === 'iOSx' ? -3 : 0;
+
+if (type === 'android') {
+  normalMarginTop = 0;
+}
 
 export default () => ({
   common: {
-    navContainer: {},
     title: {
       fontFamily: fonts.family.heading,
       fontSize: 24,
@@ -12,6 +20,9 @@ export default () => ({
       alignItems: 'center',
       justifyContent: 'space-between',
       paddingHorizontal: 20,
+      marginTop: normalMarginTop,
+      paddingRight: 20,
+      paddingTop: 10,
     },
     row: {
       flexDirection: 'row',
@@ -22,6 +33,8 @@ export default () => ({
       alignItems: 'center',
       paddingLeft: boxes.boxPadding,
       paddingRight: boxes.boxPadding,
+      marginTop: normalMarginTop,
+      paddingTop: 10,
     },
     flex: {
       flex: 1,
