+++ src/modules/BlockchainApplication/components/AddApplication/styles.js
@@ -7,12 +7,9 @@ export default function getAddApplicationStyles() {
         flex: 1,
         paddingTop: 44,
       },
-      header: {
-        marginBottom: 24
-      },
       applicationsListContainer: {
         paddingLeft: 16,
-        paddingRight: 16
+        paddingRight: 16,
       },
     },
     [themes.light]: {
