+++ src/modules/BlockchainApplication/components/ApplicationList/index.js
@@ -10,6 +10,7 @@ import ErrorIllustrationSvg from 'assets/svgs/ErrorIllustrationSvg';
 import EmptyIllustrationSvg from 'assets/svgs/EmptyIllustrationSvg';
 import ResultScreen from 'components/screens/ResultScreen';
 
+import { View } from 'react-native';
 import getBlockchainApplicationsListStyles from './styles';
 
 export default function BlockchainApplicationList({
@@ -31,21 +32,23 @@ export default function BlockchainApplicationList({
       isLoading={applications.isLoading}
       error={applications.error}
       renderData={(data) => (
-        <InfiniteScrollList
-          data={data}
-          keyExtractor={(item) => item.chainID}
-          renderItem={(item) => (
-            <Component
-              application={item}
-              navigation={navigation}
-              key={item.chainID}
-              image={item.logo.png}
-              showPinned={true}
-              onPress={() => onItemPress(item)}
-              {...props}
-            />
-          )}
-        />
+        <View style={[style?.container]}>
+          <InfiniteScrollList
+            data={data}
+            keyExtractor={(item) => item.chainID}
+            renderItem={(item) => (
+              <Component
+                application={item}
+                navigation={navigation}
+                key={item.chainID}
+                image={item.logo.png}
+                showPinned={true}
+                onPress={() => onItemPress(item)}
+                {...props}
+              />
+            )}
+          />
+        </View>
       )}
       renderLoading={() => (
         <P style={[styles.text, styles.theme.text, style?.text]}>
