+++ src/modules/Bookmark/index.js
@@ -1,4 +1,3 @@
-/* eslint-disable no-shadow */
 import React, { useState } from 'react';
 import { View, TouchableOpacity } from 'react-native';
 import { useNavigation } from '@react-navigation/native';
@@ -7,17 +6,19 @@ import { useBottomTabBarHeight } from '@react-navigation/bottom-tabs';
 
 import NavigationSafeAreaView from 'components/navigation/NavigationSafeAreaView';
 import { colors } from 'constants/styleGuide';
-import SearchBarHeader from 'components/navigation/searchBarHeader';
+import HeaderSearchBar from 'components/navigation/HeaderSearchBar';
 import Icon from 'components/shared/toolBox/icon';
 
 import { useTheme } from 'hooks/useTheme';
 import getStyles from './styles';
 import { BookmarkList } from './components';
+import { useSearch } from '../../hooks/useSearch';
 
-export default function Bookmark() {
+export default function Bookmarks() {
   const navigation = useNavigation();
 
-  const [query, setQuery] = useState('');
+  const search = useSearch();
+
   const [isSearchOpen, setIsSearchOpen] = useState(false);
 
   const tabBarHeight = useBottomTabBarHeight();
@@ -26,25 +27,25 @@ export default function Bookmark() {
     styles: getStyles(tabBarHeight),
   });
 
-  const setQueryString = (query) => setQuery(query);
+  const handleSearchTermChange = (term) => search.setTerm(term);
 
-  const onPress = (data) => {
-    navigation.navigate('Wallet', { address: data.address });
-  };
+  const handlePress = (data) => navigation.navigate('Wallet', { address: data.address });
 
   return (
     <NavigationSafeAreaView>
-      <SearchBarHeader
+      <HeaderSearchBar
         title={'Bookmarks'}
         noIcon
-        onChange={setQueryString}
-        value={query}
+        onChange={handleSearchTermChange}
+        value={search.term}
         isSearchOpen={isSearchOpen}
         setIsSearchOpen={(val) => setIsSearchOpen(val)}
       />
+
       <View style={styles.container}>
-        <BookmarkList draggable query={query} renderEmpty onPress={onPress} />
+        <BookmarkList draggable query={search.term} renderEmpty onPress={handlePress} />
       </View>
+
       <TouchableOpacity
         style={[styles.titleContainer]}
         onPress={() =>
