+++ src/modules/SendToken/components/SummaryStep/styles.js
@@ -7,8 +7,7 @@ export default function getSendTokenSummaryStepStyles() {
     common: {
       wrapper: {
         flex: 1,
-        paddingLeft: boxes.boxPadding,
-        paddingRight: boxes.boxPadding,
+        padding: boxes.boxPadding,
       },
       container: {
         flex: 1,
