+++ src/navigation/styles.js
@@ -1,10 +1,12 @@
+import { Dimensions, Platform } from 'react-native';
+
 import { colors } from 'constants/styleGuide';
-import { Dimensions } from 'react-native';
 
 export function getNavigationTabBarStyles() {
   const deviceWidth = Dimensions.get('window').width;
   const width = deviceWidth - 44;
   const left = deviceWidth / 2 - width / 2;
+  const paddingTop = Platform.OS === 'android' ? 0 : 28;
 
   return {
     position: 'absolute',
@@ -14,6 +16,6 @@ export function getNavigationTabBarStyles() {
     bottom: 24,
     borderRadius: 64,
     height: 64,
-    paddingTop: 28,
+    paddingTop,
   };
 }
